ml-engine/requirements.txt
fastapi
uvicorn
pydantic
pinecone-client
sentence-transformers
torch
numpy